Move and Clean sensor files
- optimised includes
- moved all sensor files to be under src/sensors
