+++ Embedded/src/sensors/LightSensor.hpp
