+++ Embedded/src/sensors/Temperature.hpp
